Increase the strictness of mmap() permission handling.
Convert the per-process mmap permissions value to a capability and use it
or the passed capabilities permissions to strictly limit the permissions of
the new capability.

This new capability will eventually be used to derive the returned
capability when no capability is passed in.

Reject attempts to create mapping with more permissions than the covering
capablity are now rejected at syscall entry.

To adapt libcheri to this model, add a new pass to the loading process
which protects mappings after their contents have been initalized.