Partition the 24-bit object-type space into two 23-bit spaces: those beginning with the single-bit value '0' that belong to the userspace type manager, and those beginning with the single-bit value '1' that belong to the kernel type manager.  When delegating a base type capability to userspace, authorise only use of the bottom 23 bits of the type space.
In various pieces of kernel utility code, stop always passing an
unused object-type argument during capability creation: we always
create pure memory capabilities, not sealed capabilities, and will
use separate (new) utility routines for this purpose.  This
simplifies the arguments to these various functions and macros.

Update the userspace test suite to simply expect an object-type of 0
for any capabilities found in a thread's initial register file, since
we won't offer sealed capabilities by that means -- and object types
are undefined for unsealed 128-bit capabilities.

This is the first step in allowing the kernel to implement a subset
of object types exposed to userspace.  Mysteriously, all tests pass.