Add a new per-test cheritest flag, CT_FLAG_SANDBOX, which flags tests that require sandboxing.
Add a new command-line flag "-u" to ask cheritest not to run tests that
depend on sandboxing (most useful when combined with "-a").

When "-u" is specified, do not initialise or tear down libcheri sandboxes
during test setup.

This allows testing of non-sandbox features (e.g. of CheriABI) to be
tested in isolation of kernel sandboxing support for libcheri, which has
a much broader feature footprint.