Make userspace sealing root MI and implement on RISC-V
The concept of sealing is shared by all CHERI implementations, including
Morello, yet the current sysarch(2)-based implementation for userspace
to obtain a sealing root is MIPS-specific. Instead, replace it with a
new sysctl that is shared by all CHERI implementations.

Unlike the old implementation this is now a kernel global rather than a
field attached to the proc. However, in practice the previous
implementation had the same value for the field for every single
process. If future compartmentalisation work needs to further split up
the root amongst proceses (such as the coprocess work) then the sysctl
can be made a SYSCTL_PROC and various code added to implement that as
needed, but for now that's not necessary.