cheribsdtest: Add tests to check values of RDDC_EL0/RCSP_EL0/RCTPIDR_EL0
Currently the kernel never initialises them to anything so they remain
their reset values. For RDDC_EL0 this is the almighty capability and
thus the test fails. The others reset to UNKNOWN, but in practice it
seems that means NULL on the FVP (but should absolutely not be relied
upon, even if it turns out to be true of the board).