Move /usr/libcheri to /usr/lib64c.
The NEED_COMPAT setting remains CHERI rather than being renamed to
64C.  Similarly, the rtld environment variables remain as LD_CHERI_*
rather than being renamed to LD_64C_*.

The CheriABI rtld remains ld-elf-cheri.so.1 for hybrid worlds
for now, but that should perhaps be renamed later.

This does not rename MK_COMPAT_CHERIABI to MK_LIB64C, but that could
perhaps be done as a followup commit if desired.