Revoke fully unmapped reservations
Extend the revocation framework to revoke a single map entry associated
with a fully unmapped reservation on each pass.

A quarentine list of fully unmapped reservation entries is maintained
in the vm_map and when revocation starts, one (currently the largest
one plus sufficently nearby non-adjacent neighbors) is picked for
revocation.  The revocation tests and callbacks are extended to take
start and end addresses and check each VMEM-bearing capability against
them.  At the end of the pass, the entry is removed from quarantine and
the virtual address space is once again free for use.

The key observation that makes this work is that the inheritance of
entries must be changed when they are quarantined.  We introduce a new
VM_INHERIT_QUARANTINE which we also use in place of yet another MAP_
flag to indicate that an entry is in quarantine.