test_ptrace_readtags: stack alloc & align probe region
ptrace()'s PIOD_READ_CHERI_TAGS requires that we operate on granules that are
not merely eight capabilities wide but aligned to eight capabilities as well.
calloc() does not, in general, impose sufficient alignment on its results,
though SLAB-esque allocators like jemalloc and snmalloc probably will happen to
do so.  Instead, switch to allocating the probe region on the stack and using an
attribute to align it sufficiently.