cheribsdtest: remove cheriabi_malloc_zero_size
This test asserted that zero-length capabilites were returned by
malloc(0).  While this has some nice affordances, it can't actually
work in the common case where minimum sized allocations are returned.
To understand why consider the following:

	char *a = malloc(minsize);
	char *b = malloc(0);
	assert(a + minsize == b);
	free (cheri_setbounds(a + minsize, 0)); // What is freed?

It is possible for an allocator to make this safe if there is unexposed
metadata between allocators (e.g. with dlmalloc) or if the allocator
uses dedicated slabs for zero-sized allocations.

This reverts commit e2c465d9977214a9a115b042b72080be510aad31.