Revert "mrs: Allow a program to request revocation"
The use of the malloc_revocation weak symbol seemed like a nice idea,
but it appears impractical to make it work for dynamic binaries.  Now
that revocation is on by default it's not so important to be able to
build trivial programs that enable it by default.  (Binaries that need
to disable it can be altered with elfctl or invoked via proccontrol.)

This reverts commit ffea8e0d8445085b90dfad136ad5568797b74804.