cheribsdtest: Give better error when tests unexpectedly receive a signal
We install a signal handler to catch and report signals, which then puts
the various bits of information in ccsp and does an exit(EX_SOFTWARE).
As a result, if a test receives a signal but the signal handler is able
to run and clean up like this, the test harness currently reports it as
the test failing to set a success/failure status (since the earlier
WIFSIGNALED check will see a status that exited with EX_SOFTWARE), but
we really want the root cause of that, namely the signal.